code update for occupant counts components, some minor code fixes for task list page
