+++ src/routes/TaskDetails/components/roro/tourist/group/GroupOccupants.jsx
@@ -1,15 +1,15 @@
 import React from 'react';
 import classNames from 'classnames';
 
-import Occupant from './Occupant';
+import Occupant from '../../Occupant';
 import OccupantCount from '../../OccupantCount';
 
-import { DocumentUtil, MovementUtil, PersonUtil } from '../../../../../../utils';
+import { DocumentUtil, JourneyUtil, MovementUtil, PersonUtil } from '../../../../../../utils';
 
-const Occupants = ({ version, classModifiers }) => {
+const GroupOccupants = ({ version, classModifiers }) => {
   const mode = MovementUtil.movementMode(version);
-  const journey = MovementUtil.movementJourney(version);
-  const departureTime = MovementUtil.departureTime(journey);
+  const journey = JourneyUtil.get(version);
+  const departureTime = JourneyUtil.departureTime(journey);
   const primaryTraveller = PersonUtil.get(version);
   const otherPersons = PersonUtil.getOthers(version);
   const secondaryCoTraveller = otherPersons[0] || undefined;
@@ -66,4 +66,4 @@ const Occupants = ({ version, classModifiers }) => {
   );
 };
 
-export default Occupants;
+export default GroupOccupants;
