+++ src/routes/TaskDetails/components/roro/tourist/group/TouristGroupVersionDetails.jsx
@@ -1,14 +1,18 @@
 import React from 'react';
 
-import TargetingIndicators from './TargetingIndicators';
-import Booking from './Booking';
-import Occupants from './tourist/individual/Occupants';
+import TargetingIndicators from '../../TargetingIndicators';
+import Booking from '../../Booking';
+import GroupOccupants from './GroupOccupants';
+import Document from '../../Document';
+import PrimaryTraveller from './PrimaryTraveller';
 
 const TouristVersionDetails = ({ version }) => {
   return (
     <div className="govuk-task-details-grid govuk-!-padding-top-4">
       <div className="govuk-grid-column-one-third">
-        <TargetingIndicators version={version} />
+        <TargetingIndicators version={version} classModifiers={['bottom-border-thin']} />
+        <PrimaryTraveller version={version} classModifiers={['bottom-border-thin']} />
+        <Document version={version} />
       </div>
       <div className="govuk-grid-column-one-third vertical-dotted-line__first">
         <div className="govuk-task-details__col-2">
@@ -17,7 +21,7 @@ const TouristVersionDetails = ({ version }) => {
       </div>
       <div className="govuk-grid-column-one-third vertical-dotted-line__second">
         <div className="govuk-task-details__col-3">
-          <Occupants version={version} />
+          <GroupOccupants version={version} />
         </div>
       </div>
     </div>
