+++ src/routes/TaskList/components/roro/AccompaniedMovementSection.jsx
@@ -26,7 +26,11 @@ const AccompaniedMovementSection = ({ person,
       <div className="govuk-grid-item">
         <>
           <div>
-            <EnrichmentCount labelText="Driver details" movementStats={CommonUtil.movementStats(person)} />
+            <EnrichmentCount
+              labelText="Driver details"
+              movementStats={CommonUtil.movementStats(person)}
+              classnames={['secondary-text']}
+            />
             <ul className="govuk-body-s govuk-list govuk-!-margin-bottom-0">
               <li className="govuk-!-font-weight-bold">{PersonUtil.fullname(person)}</li>
               <li>DOB: {PersonUtil.dob(person, DATE_FORMATS.SHORT)}</li>
@@ -50,7 +54,11 @@ const AccompaniedMovementSection = ({ person,
       <div className="govuk-grid-item vertical-dotted-line">
         <>
           <div>
-            <EnrichmentCount labelText="Vehicle details" movementStats={CommonUtil.movementStats(vehicle)} />
+            <EnrichmentCount
+              labelText="Vehicle details"
+              movementStats={CommonUtil.movementStats(vehicle)}
+              classnames={['secondary-text']}
+            />
             <ul className="govuk-body-s govuk-list govuk-!-margin-bottom-0">
               <li className="govuk-!-font-weight-bold">{VehicleUtil.registration(vehicle)}</li>
               <li>{VehicleUtil.colour(vehicle)}</li>
@@ -62,7 +70,7 @@ const AccompaniedMovementSection = ({ person,
             <EnrichmentCount
               labelText="Trailer details"
               movementStats={CommonUtil.movementStats(trailer)}
-              classnames="govuk-!-margin-top-3"
+              classnames={['secondary-text', 'govuk-!-margin-top-3']}
             />
             <ul className="govuk-body-s govuk-list govuk-!-margin-bottom-0">
               <li className="govuk-!-font-weight-bold">{TrailerUtil.registration(trailer)}</li>
@@ -78,6 +86,7 @@ const AccompaniedMovementSection = ({ person,
             <EnrichmentCount
               labelText="Haulier details"
               movementStats={CommonUtil.movementStats(haulier)}
+              classnames={['secondary-text']}
             />
             <ul className="govuk-body-s govuk-list govuk-!-margin-bottom-0">
               <li className="govuk-!-font-weight-bold">{HaulierUtil.name(haulier)}</li>
@@ -87,7 +96,7 @@ const AccompaniedMovementSection = ({ person,
             <EnrichmentCount
               labelText="Account details"
               movementStats={CommonUtil.movementStats(account)}
-              classnames="govuk-!-margin-top-3"
+              classnames={['secondary-text', 'govuk-!-margin-top-3']}
             />
             <ul className="govuk-body-s govuk-list govuk-!-margin-bottom-0">
               <li className="govuk-!-font-weight-bold">{AccountUtil.name(account)}</li>
