+++ src/routes/TaskList/components/roro/UnaccompaniedMovementSection.jsx
@@ -25,6 +25,7 @@ const UnaccompaniedMovementSection = ({ trailer,
             <EnrichmentCount
               labelText="Trailer details"
               movementStats={CommonUtil.movementStats(trailer)}
+              classnames={['secondary-text']}
             />
             <ul className="govuk-body-s govuk-list govuk-!-margin-bottom-0">
               <li className="govuk-!-font-weight-bold">{TrailerUtil.registration(trailer)}</li>
@@ -40,6 +41,7 @@ const UnaccompaniedMovementSection = ({ trailer,
             <EnrichmentCount
               labelText="Haulier details"
               movementStats={CommonUtil.movementStats(haulier)}
+              classnames={['secondary-text']}
             />
             <ul className="govuk-body-s govuk-list govuk-!-margin-bottom-0">
               <li className="govuk-!-font-weight-bold">{HaulierUtil.name(haulier)}</li>
@@ -49,7 +51,7 @@ const UnaccompaniedMovementSection = ({ trailer,
             <EnrichmentCount
               labelText="Account details"
               movementStats={CommonUtil.movementStats(account)}
-              classnames="govuk-!-margin-top-3"
+              classnames={['secondary-text', 'govuk-!-margin-top-3']}
             />
             <ul className="govuk-body-s govuk-list govuk-!-margin-bottom-0">
               <li className="govuk-!-font-weight-bold">{AccountUtil.name(account)}</li>
